// grad_project-b4_search/src/types.ts
export interface Itinerary {
  title: string;
  dayNumber: number;
  description: string;
  startTime: string;
  endTime: string;
  location: string;
  activities: string[];
  includesBreakfast: boolean;
  includesLunch: boolean;
  includesDinner: boolean;
  notes: string;
}

export interface Tour {
  id: number;
  title: string;
  description: string;
  price: number;
  startDate: string;
  endDate: string;
  creationDate?: string;
  availableSeats: number;
  departurePoint: string;
  departurePointLat?: number;
  departurePointLng?: number;
  destinationCity: string;
  destinationCityLat?: number;
  destinationCityLng?: number;
  transportationType: string;
  coverImageUrl: string;
  amenities: string[];
  companyId?: number;
  companyName?: string;
  profileImageUrl?: string;
  companyProfileImageUrl?: string;
  imageUrls: string[];
  image?: string;
  tags?: string[];
  rating?: number;
  itineraries?: Itinerary[];
  categoryId?: number;
}
export interface Event {
  id: number;
  title: string;
  category: string;
  dates?: string | null;
  date: string;
  location: string;
  link: string;
  image: string;
  map_Link?: string | null;
  type?: string | null;
  description?: string | null;
  url?: string | null;
}

export interface Category {
  id: number;
  categoryName: string;
}